refactor: 회원 가입 페이지 UI 개선
